[libromdata] EXE_NE: Explicitly cast len to size_t for vhvc::span<>.
This is a workaround for the lack of SFINAE support in MSVC 2015.
It attempts to use the span<It, End> overload instead of the
span<It, size_type> overload.

+++ src/libromdata/Other/EXE_NE.cpp
@@ -583,7 +583,7 @@ int EXEPrivate::addFields_NE_Entry(void)
 			uint8_t len = *p++;
 			if (p + len + 2 >= end) // next length byte >= end
 				return -ENOENT;
-			vhvc::span<const char> name(reinterpret_cast<const char *>(p), len);
+			vhvc::span<const char> name(reinterpret_cast<const char *>(p), static_cast<size_t>(len));
 			uint16_t ordinal = p[len] | p[len+1]<<8;
 
 			// binary search for the ordinal
